requirements.txt
cffi==2.0.0
colorama==0.4.6
flatbuffers==25.12.19
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.2
packaging==26.0
protobuf==7.34.0
pycparser==3.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyqtgraph==0.14.0
PyYAML==6.0.3
scipy==1.17.1
soundfile==0.13.1
sympy==1.14.0
